# Async clearing config index