--- a/pilot/pkg/model/xds_cache_test.go
+++ b/pilot/pkg/model/xds_cache_test.go
@@ -19,6 +19,7 @@ import (
 	"time"
 
 	discovery "github.com/envoyproxy/go-control-plane/envoy/service/discovery/v3"
+	"k8s.io/apimachinery/pkg/util/wait"
 
 	"istio.io/istio/pilot/pkg/features"
 	"istio.io/istio/pkg/config/schema/kind"
@@ -101,6 +102,7 @@ func TestAddTwoEntries(t *testing.T) {
 }
 
 func TestCleanIndexesOnAddExistant(t *testing.T) {
+	test.SetForTest(t, &features.XDSCacheIndexClearInterval, 5*time.Millisecond)
 	zeroTime := time.Time{}
 	res := &discovery.Resource{Name: "test"}
 	req := &PushRequest{Start: zeroTime.Add(time.Duration(1))}
@@ -139,7 +141,12 @@ func TestCleanIndexesOnAddExistant(t *testing.T) {
 	c.Add(&secondEntry, req, res)
 
 	assert.Equal(t, cache.store.Len(), 1)
-	assert.Equal(t, len(cache.configIndex), 1)
+	err := wait.Poll(2*time.Millisecond, 100*time.Millisecond, func() (bool, error) {
+		cache.mu.RLock()
+		defer cache.mu.RUnlock()
+		return len(cache.configIndex) == 1, nil
+	})
+	assert.Equal(t, err, nil)
 	assert.Equal(t, cache.configIndex, map[ConfigHash]sets.String{
 		ConfigKey{Kind: kind.DestinationRule, Name: "name", Namespace: "namespace"}.HashCode(): sets.New(secondEntry.key),
 	})
@@ -147,6 +154,7 @@ func TestCleanIndexesOnAddExistant(t *testing.T) {
 
 func TestCleanIndexesOnEvict(t *testing.T) {
 	test.SetForTest(t, &features.XDSCacheMaxSize, 1)
+	test.SetForTest(t, &features.XDSCacheIndexClearInterval, 5*time.Millisecond)
 	zeroTime := time.Time{}
 	res := &discovery.Resource{Name: "test"}
 	req := &PushRequest{Start: zeroTime.Add(time.Duration(1))}
@@ -161,6 +169,9 @@ func TestCleanIndexesOnEvict(t *testing.T) {
 
 	c := NewXdsCache()
 	cache := c.(*lruCache)
+	stop := make(chan struct{})
+	defer close(stop)
+	cache.Run(stop)
 
 	assert.Equal(t, cache.store.Len(), 0)
 	assert.Equal(t, cache.configIndex, map[ConfigHash]sets.String{})
@@ -189,7 +200,12 @@ func TestCleanIndexesOnEvict(t *testing.T) {
 	c.Add(&secondEntry, req, res)
 
 	assert.Equal(t, cache.store.Len(), 1)
-	assert.Equal(t, len(cache.configIndex), 2)
+	err := wait.Poll(2*time.Millisecond, 100*time.Millisecond, func() (bool, error) {
+		cache.mu.RLock()
+		defer cache.mu.RUnlock()
+		return len(cache.configIndex) == 2, nil
+	})
+	assert.Equal(t, err, nil)
 	assert.Equal(t, cache.configIndex, map[ConfigHash]sets.String{
 		ConfigKey{Kind: kind.Service, Name: "name", Namespace: "namespace"}.HashCode():     sets.New(secondEntry.key),
 		ConfigKey{Kind: kind.EnvoyFilter, Name: "name", Namespace: "namespace"}.HashCode(): sets.New(secondEntry.key),
@@ -198,6 +214,7 @@ func TestCleanIndexesOnEvict(t *testing.T) {
 
 func TestCleanIndexesOnCacheClear(t *testing.T) {
 	test.SetForTest(t, &features.XDSCacheMaxSize, 10)
+	test.SetForTest(t, &features.XDSCacheIndexClearInterval, 5*time.Millisecond)
 	zeroTime := time.Time{}
 	res := &discovery.Resource{Name: "test"}
 	req1 := &PushRequest{Start: zeroTime.Add(time.Duration(1))}
@@ -260,7 +277,10 @@ func TestCleanIndexesOnCacheClear(t *testing.T) {
 	cache.Clear(sets.Set[ConfigKey]{{Kind: kind.DestinationRule, Name: "name", Namespace: "namespace"}: {}})
 
 	assert.Equal(t, cache.store.Len(), 1)
-	assert.Equal(t, len(cache.configIndex), 3)
+	err := wait.Poll(2*time.Millisecond, 100*time.Millisecond, func() (done bool, err error) {
+		return len(cache.configIndex) == 3, nil
+	})
+	assert.Equal(t, err, nil)
 	assert.Equal(t, cache.configIndex, map[ConfigHash]sets.String{
 		ConfigKey{Kind: kind.Service, Name: "name", Namespace: "namespace"}.HashCode():     sets.New(secondEntry.key),
 		ConfigKey{Kind: kind.EnvoyFilter, Name: "name", Namespace: "namespace"}.HashCode(): sets.New(secondEntry.key),
@@ -271,7 +291,10 @@ func TestCleanIndexesOnCacheClear(t *testing.T) {
 	c.Add(&firstEntry, &PushRequest{Start: zeroTime.Add(time.Duration(3))}, res)
 
 	assert.Equal(t, cache.store.Len(), 2)
-	assert.Equal(t, len(cache.configIndex), 5)
+	err = wait.Poll(2*time.Millisecond, 100*time.Millisecond, func() (done bool, err error) {
+		return len(cache.configIndex) == 5, nil
+	})
+	assert.Equal(t, err, nil)
 	assert.Equal(t, cache.configIndex, map[ConfigHash]sets.String{
 		ConfigKey{Kind: kind.Service, Name: "name", Namespace: "namespace"}.HashCode():         sets.New(firstEntry.key, secondEntry.key),
 		ConfigKey{Kind: kind.DestinationRule, Name: "name", Namespace: "namespace"}.HashCode(): sets.New(firstEntry.key),
@@ -284,7 +307,10 @@ func TestCleanIndexesOnCacheClear(t *testing.T) {
 	cache.Clear(sets.Set[ConfigKey]{{Kind: kind.EnvoyFilter, Name: "name", Namespace: "namespace"}: {}})
 
 	assert.Equal(t, cache.store.Len(), 1)
-	assert.Equal(t, len(cache.configIndex), 3)
+	err = wait.Poll(2*time.Millisecond, 100*time.Millisecond, func() (done bool, err error) {
+		return len(cache.configIndex) == 3, nil
+	})
+	assert.Equal(t, err, nil)
 	assert.Equal(t, cache.configIndex, map[ConfigHash]sets.String{
 		ConfigKey{Kind: kind.Service, Name: "name", Namespace: "namespace"}.HashCode():         sets.New(firstEntry.key),
 		ConfigKey{Kind: kind.DestinationRule, Name: "name", Namespace: "namespace"}.HashCode(): sets.New(firstEntry.key),
@@ -295,7 +321,11 @@ func TestCleanIndexesOnCacheClear(t *testing.T) {
 	c.Add(&secondEntry, &PushRequest{Start: zeroTime.Add(time.Duration(4))}, res)
 
 	assert.Equal(t, cache.store.Len(), 2)
-	assert.Equal(t, len(cache.configIndex), 5)
+
+	err = wait.Poll(2*time.Millisecond, 100*time.Millisecond, func() (done bool, err error) {
+		return len(cache.configIndex) == 5, nil
+	})
+	assert.Equal(t, err, nil)
 	assert.Equal(t, cache.configIndex, map[ConfigHash]sets.String{
 		ConfigKey{Kind: kind.Service, Name: "name", Namespace: "namespace"}.HashCode():         sets.New(firstEntry.key, secondEntry.key),
 		ConfigKey{Kind: kind.DestinationRule, Name: "name", Namespace: "namespace"}.HashCode(): sets.New(firstEntry.key),
@@ -307,7 +337,10 @@ func TestCleanIndexesOnCacheClear(t *testing.T) {
 	// clear only Service dependencies, should clear both firstEntry and secondEntry references
 	cache.Clear(sets.Set[ConfigKey]{{Kind: kind.Service, Name: "name", Namespace: "namespace"}: {}})
 
-	assert.Equal(t, len(cache.configIndex), 0)
+	err = wait.Poll(2*time.Millisecond, 100*time.Millisecond, func() (done bool, err error) {
+		return len(cache.configIndex) == 0, nil
+	})
+	assert.Equal(t, err, nil)
 	assert.Equal(t, cache.store.Len(), 0)
 	assert.Equal(t, cache.configIndex, map[ConfigHash]sets.String{})
 }