**Please provide a description of this PR:**


For https://github.com/istio/istio/issues/43431