@hzxuzhonghu 2023-02-22
```
root@kurator-linux-0001:~/go/src/istio.io/istio/pilot/pkg/xds# benchstat old.txt new.txt 
goos: linux
goarch: amd64
pkg: istio.io/istio/pilot/pkg/xds
cpu: Intel(R) Xeon(R) Gold 6278C CPU @ 2.60GHz
                        │   old.txt    │               new.txt               │
                        │    sec/op    │    sec/op     vs base               │
Cache/key-4                8.737µ ± 1%   8.759µ ±  1%        ~ (p=0.818 n=6)
Cache/insert-4             182.7µ ± 3%   191.6µ ±  9%        ~ (p=0.240 n=6)
Cache/insert_same_key-4   137.91µ ± 5%   97.25µ ± 14%  -29.48% (p=0.002 n=6)
Cache/get-4                8.869µ ± 1%   8.867µ ±  1%        ~ (p=0.818 n=6)
Cache/insert_and_get-4     192.4µ ± 3%   198.0µ ±  4%        ~ (p=0.093 n=6)
geomean                    51.87µ        49.14µ         -5.27%
```