@istio-testing 2023-02-22
@hzxuzhonghu: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
release-notes_istio | c3d60a07defdfff4d068d5a6d727e9be7f74a02f | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/43524/release-notes_istio/1628242425988255744) | true | `/test release-notes`
lint_istio | c3d60a07defdfff4d068d5a6d727e9be7f74a02f | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/43524/lint_istio/1628242423782051840) | true | `/test lint`
unit-tests-arm64_istio | c3d60a07defdfff4d068d5a6d727e9be7f74a02f | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/43524/unit-tests-arm64_istio/1628242413753470976) | true | `/test unit-tests-arm64`
unit-tests_istio | c3d60a07defdfff4d068d5a6d727e9be7f74a02f | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/43524/unit-tests_istio/1628242413715722240) | true | `/test unit-tests`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->