@howardjohn 2023-02-23
This feels like a bit of complexity/risk for minimal gain - only 30% in a specific case, but not generalized use cases. I wonder if we should tackle the real root of the issue which I suspect is that we have to use a Lock instead of a rwlock for the LRU. https://github.com/istio/istio/issues/40744 shows not much info - only a mutex profile. We should have them show more info on profiles if we are going to hyper optimize for their case.

Maybe we reconsider use of LRU entirely to avoid locking?

BTW https://dgraph.io/blog/post/caching-in-go/ is fairly interesting. I took a stab at integrating their approach into our cache but it doesn't work to well for the same reason we have issues here with configIndex needing its own lock.

IDK, I feel like we are going down a rabbit hole when its not clear these are actually providing real world benefits or just microbenchmarks. I don't want to block things too much though if we think this will have benefits... Even with this change we still lock just as much as before (in terms of raw number of locks, we spend a bit less time under the lock). Just thinking if we may want to consider bigger changes to really remove lock usage?


wdyt @ramaraochavali @stevenctl ? I am fine with this if others want to proceed