@hzxuzhonghu 2023-02-23
I agree the test is far from covering all cases. But I would suggest we look this from a different side, does this patch bring any bad or regression while it brings improvement in explicit case?