@howardjohn 2023-02-23
Thinking about it a bit more I am actually less concerned with this than https://github.com/istio/istio/pull/43428. Are they coupled?

That one may be fine too, I am still doing research into the safety part (https://github.com/istio/istio/pull/43428#discussion_r1113320799) .I think you may be right, just want to verify.

> does this patch bring any bad or regression while it brings improvement in explicit case?

Well it is a bit more complex... but not too bad.