@hzxuzhonghu 2023-02-23
It has some dependency since the clearing index procedure includes both typeIndex and configIndex. If typeIndex removed, we only need to consider if there is any side-effect of config index